3eme tentative de correction de cast

+++ Triton_Layers/SSA.py
@@ -246,7 +246,7 @@ class SSAMultiHeadAttention(nn.Module):
         log_bias = self.get_log_pe_bias_cross(N, N_k, Q.device)  # (L, L)
         
         # Broadcast sur batch et heads: (L, L) -> (1, 1, L, L)
-        log_bias = log_bias.unsqueeze(0).unsqueeze(0)
+        log_bias = log_bias.unsqueeze(0).unsqueeze(0).to(dtype=Q.dtype)
         attn_output = attn_output + log_bias  # (B, n_heads, L, L)
         
         if attention_mask is not None:
